Modified jwtsettings and applicationServiceConfigurationExtention.

+++ luxe/luxe.Server/Infrastructure/Configurations/JwtSettings.cs
@@ -6,6 +6,6 @@
         public string Issuer { get; set; } = string.Empty;
         public string Audience { get; set; } = string.Empty;
         public int MinutesToExpiration { get; set; }
-        public int RefreshTokenExpirationDays { get; set; }
+        public double RefreshTokenExpirationDays { get; set; }
     }
 }
